# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GoMindscout API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: deleteMe
      tags: [users]
      summary: Delete current user account and all associated data
      responses:
        "204":
          description: Account deleted
        "401":
          description: Unauthorized

  # ── Categories ─────────────────────────────────────────────────────────────
...
  /mentors/{mentorId}/packages:
    get:
      operationId: listMentorPackages
      tags: [packages]
      summary: List packages for a mentor
      parameters:
        - name: mentorId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
  /packages:
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a package for own mentor profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{packageId}:
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package
      parameters:
        - name: packageId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Deactivate a package
      parameters:
        - name: packageId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Bookings ───────────────────────────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews (optionally filter by bookingId)
      parameters:
        - name: bookingId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a completed booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /mentors/{mentorId}/reviews:
    get:
      operationId: listMentorReviews
      tags: [reviews]
      summary: Get reviews for a mentor
      parameters:
        - name: mentorId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # ── Dashboard stats (wow endpoints) ────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, email, role, createdAt]
      properties:
        id: { type: integer }
        clerkId: { type: string }
        email: { type: string }
        fullName: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        role: { type: string, enum: [mentee, mentor, admin] }
        createdAt: { type: string }
    UserProfileUpdate:
      type: object
      properties:
        fullName: { type: string }
        avatarUrl: { type: string }
...
    Package:
      type: object
      required: [id, mentorId, title, type, price, isActive]
      properties:
        id: { type: integer }
        mentorId: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        type: { type: string, enum: [video_30, video_60] }
        durationMinutes: { type: ["integer", "null"] }
        price: { type: number }
        isActive: { type: boolean }
    PackageInput:
      type: object
      required: [title, type, price]
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string, enum: [video_30, video_60] }
        durationMinutes: { type: integer }
        price: { type: number }
    PackageUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        price: { type: number }
        isActive: { type: boolean }
...
    Review:
      type: object
      required: [id, bookingId, mentorId, menteeId, rating, createdAt]
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        mentorId: { type: integer }
        menteeId: { type: integer }
        rating: { type: integer }
        punctualityRating: { type: ["integer", "null"] }
        communicationRating: { type: ["integer", "null"] }
        valueRating: { type: ["integer", "null"] }
        comment: { type: ["string", "null"] }
        menteeName: { type: ["string", "null"] }
        menteeAvatarUrl: { type: ["string", "null"] }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [bookingId, rating]
      properties:
        bookingId: { type: integer }
        rating: { type: integer }
        punctualityRating: { type: integer }
        communicationRating: { type: integer }
        valueRating: { type: integer }
        comment: { type: string }
...